backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
httpx==0.27.0
python-dotenv==1.0.1
python-multipart==0.0.9
Pillow==10.3.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4